+++ EasyLife/PlanItemContext.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+enum PlanItemContext {
+    case existing(item: TodoItem)
+    case new(item: TodoItem, context: DataContexting)
+}