+++ EasyLife/ProjectsRepository.swift
@@ -15,19 +15,19 @@ protocol ProjectsRepositoring: Mockable {
 }
 
 final class ProjectsRepository: ProjectsRepositoring {
-    private let dataManager: DataManaging
+    private let dataProvider: DataContextProviding
     private let priorityPredicate = NSPredicate(format: "%K != %d",
                                                 argumentArray: ["priority", Project.defaultPriority])
     private let otherPredicate = NSPredicate(format: "%K = %d", argumentArray: ["priority", Project.defaultPriority])
 
-    init(dataManager: DataManaging) {
-        self.dataManager = dataManager
+    init(dataProvider: DataContextProviding) {
+        self.dataProvider = dataProvider
     }
 
     func delete(project: Project) -> Async<Void> {
         return Async { completion in
             async({
-                let context = self.dataManager.mainContext()
+                let context = self.dataProvider.mainContext()
                 context.delete(project)
                 _ = try await(context.save())
                 completion(.success(()))
@@ -40,7 +40,7 @@ final class ProjectsRepository: ProjectsRepositoring {
     func addProject(name: String) -> Async<Project> {
         return Async { completion in
             async({
-                let context = self.dataManager.mainContext()
+                let context = self.dataProvider.mainContext()
                 let project = context.insert(entityClass: Project.self)
                 project.name = name
                 _ = try await(context.save())
@@ -55,7 +55,7 @@ final class ProjectsRepository: ProjectsRepositoring {
         return Async { completion in
             async({
                 project.name = name
-                _ = try await(self.dataManager.mainContext().save())
+                _ = try await(self.dataProvider.mainContext().save())
                 completion(.success(()))
             }, onError: { error in
                 completion(.failure(error))
@@ -67,7 +67,7 @@ final class ProjectsRepository: ProjectsRepositoring {
         assert(max > 0)
         return Async { completion in
             async({
-                let context = self.dataManager.mainContext()
+                let context = self.dataProvider.mainContext()
                 let projects = try await(context.fetch(
                     entityClass: Project.self,
                     sortBy: nil,
@@ -89,7 +89,7 @@ final class ProjectsRepository: ProjectsRepositoring {
     func deprioritize(project: Project) -> Async<Void> {
         return Async { completion in
             async({
-                let context = self.dataManager.mainContext()
+                let context = self.dataProvider.mainContext()
                 let predicate = NSPredicate(format: "%K > %d",
                                             argumentArray: ["priority", project.priority])
                 let projects = try await(context.fetch(
@@ -110,7 +110,7 @@ final class ProjectsRepository: ProjectsRepositoring {
     func prioritise(_ projectA: Project, above projectB: Project) -> Async<Void> {
         return Async { completion in
             async({
-                let context = self.dataManager.mainContext()
+                let context = self.dataProvider.mainContext()
                 let projectADestinationPriority = projectB.priority
                 let predicate = NSPredicate(
                     format: "%K >= %d AND %K < %d",
@@ -134,7 +134,7 @@ final class ProjectsRepository: ProjectsRepositoring {
     func prioritise(_ projectA: Project, below projectB: Project) -> Async<Void> {
         return Async { completion in
             async({
-                let context = self.dataManager.mainContext()
+                let context = self.dataProvider.mainContext()
                 let projectADestinationPriority = projectB.priority
                 let predicate = NSPredicate(
                     format: "%K <= %d AND %K != %d",
@@ -158,7 +158,7 @@ final class ProjectsRepository: ProjectsRepositoring {
     func fetchPrioritizedProjects() -> Async<[Project]> {
         return Async { completion in
             async({
-                let context = self.dataManager.mainContext()
+                let context = self.dataProvider.mainContext()
                 let projects = try await(context.fetch(
                     entityClass: Project.self,
                     sortBy: DataSort(sortDescriptor: [NSSortDescriptor(key: "priority", ascending: true)]),
@@ -174,7 +174,7 @@ final class ProjectsRepository: ProjectsRepositoring {
     func fetchOtherProjects() -> Async<[Project]> {
         return Async { completion in
             async({
-                let context = self.dataManager.mainContext()
+                let context = self.dataProvider.mainContext()
                 let projects = try await(context.fetch(
                     entityClass: Project.self,
                     sortBy: DataSort(sortDescriptor: [NSSortDescriptor(key: "name", ascending: true)]),