+++ EasyLifeTests/ArchiveTests.swift
@@ -180,8 +180,8 @@ final class ArchiveTests: XCTestCase {
         // test
         waitAsync(delay: 0.5) { completion in
             async({
-                let items = try await(self.env.dataManager.fetch(entityClass: TodoItem.self, sortBy: nil,
-                                                                 context: .main, predicate: nil))
+                let items = try await(self.env.dataProvider.fetch(entityClass: TodoItem.self, sortBy: nil,
+                                                                  context: .main, predicate: nil))
                 XCTAssertEqual(items.count, 0)
                 completion()
             }, onError: { error in