+++ EasyLifeTests/ItemDetailTests.swift
@@ -54,8 +54,8 @@ final class ItemDetailTests: XCTestCase {
         // test
         waitAsync { completion in
             async({
-                let items = try await(self.env.dataManager.fetch(entityClass: TodoItem.self, sortBy: nil,
-                                                                 context: .main, predicate: nil))
+                let items = try await(self.env.dataProvider.fetch(entityClass: TodoItem.self, sortBy: nil,
+                                                                  context: .main, predicate: nil))
                 XCTAssertEqual(items.count, 1)
                 completion()
             }, onError: { error in
@@ -77,8 +77,8 @@ final class ItemDetailTests: XCTestCase {
         // test
         waitAsync(delay: 0.5) { completion in
             async({
-                let items = try await(self.env.dataManager.fetch(entityClass: TodoItem.self, sortBy: nil,
-                                                                 context: .main, predicate: nil))
+                let items = try await(self.env.dataProvider.fetch(entityClass: TodoItem.self, sortBy: nil,
+                                                                  context: .main, predicate: nil))
                 XCTAssertEqual(items.count, 1)
                 completion()
             }, onError: { error in
@@ -104,8 +104,8 @@ final class ItemDetailTests: XCTestCase {
         // test
         waitAsync(delay: 0.5) { completion in
             async({
-                let items = try await(self.env.dataManager.fetch(entityClass: TodoItem.self, sortBy: nil,
-                                                                 context: .main, predicate: nil))
+                let items = try await(self.env.dataProvider.fetch(entityClass: TodoItem.self, sortBy: nil,
+                                                                  context: .main, predicate: nil))
                 XCTAssertEqual(items.first?.project, project)
                 XCTAssertEqual(items.count, 1)
                 completion()
@@ -149,8 +149,8 @@ final class ItemDetailTests: XCTestCase {
         // test
         waitAsync(delay: 0.5) { completion in
             async({
-                let items = try await(self.env.dataManager.fetch(entityClass: TodoItem.self, sortBy: nil,
-                                                                 context: .main, predicate: nil))
+                let items = try await(self.env.dataProvider.fetch(entityClass: TodoItem.self, sortBy: nil,
+                                                                  context: .main, predicate: nil))
                 XCTAssertEqual(items.count, 0)
                 completion()
             }, onError: { error in