+++ EasyLifeTests/PlanTests.swift
@@ -470,8 +470,8 @@ final class PlanTests: XCTestCase {
 
         // test
         waitAsync(delay: 0.5, queue: .asyncAwait) { completion in
-            let items = try? await(self.env.dataManager.fetch(entityClass: TodoItem.self, sortBy: nil, context: .main,
-                                                              predicate: nil))
+            let items = try? await(self.env.dataProvider.fetch(entityClass: TodoItem.self, sortBy: nil, context: .main,
+                                                               predicate: nil))
             XCTAssertEqual(items?.count, 0)
             completion()
         }
@@ -494,8 +494,8 @@ final class PlanTests: XCTestCase {
 
         // test
         waitAsync(delay: 0.5, queue: .asyncAwait) { completion in
-            let items = try? await(self.env.dataManager.fetch(entityClass: TodoItem.self, sortBy: nil, context: .main,
-                                                              predicate: nil))
+            let items = try? await(self.env.dataProvider.fetch(entityClass: TodoItem.self, sortBy: nil, context: .main,
+                                                               predicate: nil))
             XCTAssertNotNil(items?.first(where: { $0.date == date && $0.repeatState == RepeatState.none }))
             XCTAssertNotNil(items?.first(where: { ($0.date ?? date) > date
                 && $0.repeatState == .daily && ($0.blockedBy?.count ?? 0) == 0 }))