+++ EasyLifeTests/ProjectsTests.swift
@@ -181,8 +181,8 @@ final class ProjectsTests: XCTestCase {
         // test
         waitAsync(delay: 0.5) { completion in
             async({
-                let items = try await(self.env.dataManager.fetch(entityClass: Project.self, sortBy: nil,
-                                                                 context: .main, predicate: nil))
+                let items = try await(self.env.dataProvider.fetch(entityClass: Project.self, sortBy: nil,
+                                                                  context: .main, predicate: nil))
                 XCTAssertEqual(items.count, 1)
                 completion()
             }, onError: { error in
@@ -397,8 +397,8 @@ final class ProjectsTests: XCTestCase {
         // test
         waitAsync(delay: 0.5) { completion in
             async({
-                let items = try await(self.env.dataManager.fetch(entityClass: Project.self, sortBy: nil,
-                                                                 context: .main, predicate: nil))
+                let items = try await(self.env.dataProvider.fetch(entityClass: Project.self, sortBy: nil,
+                                                                  context: .main, predicate: nil))
                 XCTAssertEqual(items.count, 0)
                 completion()
             }, onError: { error in